+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -742,6 +742,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       public void testOverridesAreTypeChecked() {
         run("OverridesAreTypeChecked.kt");
       }
+
+      @Test
+      @TestMetadata("ParentDuplicateBindingsUsedOnlyByChild.kt")
+      public void testParentDuplicateBindingsUsedOnlyByChild() {
+        run("ParentDuplicateBindingsUsedOnlyByChild.kt");
+      }
     }
 
     @Nested